[test] Move pytest fixtures to its own module
Again, to keep things organized.
